webapi/webapi-webrtc-w3c-tests/webrtc/RTCDTMFToneChangeEvent_tone_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu,Yun <yunx.liu@intel.com>

-->

<meta charset="utf-8">
<title>WebRTC Test: RTCDTMFToneChangeEvent - tone basic</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/webrtc.html#rtcdtmftonechangeevent">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>
var t = async_test("Check if the readonly attribute tone of RTCDTMFToneChangeEvent exists and type of string");

var configuration = {"iceServers": [{"url": "stun:stun.example.org"}]};
var pc1;
var pc2;
var localStream;
if (typeof RTCPeerConnection != "undefined") {
  pc1 = new RTCPeerConnection(configuration);
  pc2 = new RTCPeerConnection(configuration);
} else {
  pc1 = new webkitRTCPeerConnection(configuration);
  pc2 = new RTCPeerConnection(configuration);
}

pc1.onicecandidate = t.step_func(function(evt) {
  if(evt.candidate) {
    pc2.addIceCandidate(event.candidate);
  }
});

pc2.onicecandidate = t.step_func(function(evt) {
  if(evt.candidate) {
    pc1.addIceCandidate(event.candidate);
  }
});

pc2.onaddstream = t.step_func(function(e) {
  var dtmfsender = pc1.createDTMFSender(localStream.getAudioTracks()[0]);
  dtmfsender.ontonechange = t.step_func(function(event) {
    assert_true("tone" in event, "tone attribute in RTCDTMFToneChangeEvent");
    assert_equals(Object.prototype.toString.call(event.tone), "[object String]", "tone attribute of type");
    assert_readonly(event, "tone", "expect attribute tone cannot be changed but");
    t.done();
  });
  setTimeout(function() {
    dtmfsender.insertDTMF("1234567890#*");
  }, 100);
});

t.step(function () {
  if (typeof (navigator.getUserMedia) != "undefined") {
    navigator.getUserMedia({audio:true, video:true}, gotStream, error);
  } else {
    navigator.webkitGetUserMedia({audio:true, video:true}, gotStream, error);
  }
});

function gotStream(stream) {
  localStream = stream;
  pc1.addStream(localStream);
  pc1.createOffer(requestSucceeded1, requestFailed1);
}

function requestSucceeded1 (desc) {
  pc1.setLocalDescription(desc);
  pc2.setRemoteDescription(desc);
  pc2.createAnswer(requestSucceeded2, requestFailed2);
}

function requestFailed1 () {
  t.step(function () {
    assert_unreached("requestFailed1 was called");
  });
  t.done();
}

function requestSucceeded2(desc) {
  pc2.setLocalDescription(desc);
  pc1.setRemoteDescription(desc);
}

function requestFailed2 () {
  t.step(function () {
    assert_unreached("requestFailed2 was called");
  });
  t.done();
}

function error() {
  t.step(function () {
    assert_unreached("Stream generation failed");
  });
  t.done();
}

</script>
